requirement.txt
appdirs==1.4.3
Babel==2.7.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
bs4==0.0.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.14
fake-useragent==0.1.11
fdfgen==0.16.1
Flask==1.1.1
Flask-BabelEx==0.9.3
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
html5lib==1.0.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
lxml==4.4.2
MarkupSafe==1.1.1
matplotlib==3.1.2
numpy==1.18.0
pandas==0.25.3
parse==1.14.0
passlib==1.7.2
pdfkit==0.6.1
Pillow==6.2.1
pycparser==2.19
pyee==6.0.0
pyparsing==2.4.6
PyPDF2==1.26.0
pypdftk==0.4
pyppeteer==0.0.25
pyquery==1.4.1
pytesseract==0.3.1
python-dateutil==2.8.1
pytz==2019.3
reportlab==3.5.32
requests==2.22.0
requests-html==0.10.0
six==1.13.0
soupsieve==1.9.5
speaklater==1.3
SQLAlchemy==1.3.12
tqdm==4.41.1
urllib3==1.25.7
w3lib==1.21.0
webencodings==0.5.1
websockets==8.1
Werkzeug==0.16.0
wkhtmltopdf==0.2
WTForms==2.2.1
xlrd==1.2.0